Add trace log and debug option

+++ R/irace_output_view.R
@@ -103,12 +103,19 @@ IraceOutputView <- R6::R6Class(
         store$iraceAlive()
 
         if (!store$iraceProcess$is_alive()) {
+          log_info("Irace end.")
           enable(id = "scenarioPicker")
 
-          unlink(pkg$tempFolder, recursive = TRUE, force = TRUE)
+          if (!get_option("debug", FALSE)) {
+            unlink(pkg$tempFolder, recursive = TRUE, force = TRUE)
+          }
 
-          store$iraceProcess$poll_io(1000)
-          error <- store$iraceProcess$read_all_error_lines()
+          store$iraceProcess$poll_io(1500)
+          error <- store$iraceProcess$read_error_lines()
+
+          if (get_option("debug", FALSE)) {
+            log_error("Optional: {error}")
+          }
 
           log <- gsub('"', "", store$pg$get_irace_option("logFile"))
